backend/requirements.txt
fastapi
uvicorn
pandas
openai
sqlalchemy
psycopg2-binary
cryptography
requests
python-dotenv